splittin it up
